+++ packages/uniforms-bootstrap4/test/combined.js
@@ -158,7 +158,7 @@ describe('Everything', () => {
     it('works (NumField, invalid)', async () => {
         expect(wrapper.find('#x00').props()).to.have.property('value', 0);
         expect(wrapper.find('#x00').simulate('change', {target: {value: 'invalid'}})).to.be.ok;
-        expect(wrapper.find('#x00').props()).to.have.property('value', 0);
+        expect(wrapper.find('#x00').props()).to.have.property('value', '');
 
         await new Promise(resolve => setTimeout(resolve, 5));
 
@@ -283,7 +283,7 @@ describe('Everything', () => {
     it('works (NumField, decimal, nullable)', async () => {
         expect(wrapper.find('#x22').props()).to.have.property('value', 0);
         expect(wrapper.find('#x22').simulate('change', {target: {value: ''}})).to.be.ok;
-        expect(wrapper.find('#x22').props()).to.have.property('value', 0);
+        expect(wrapper.find('#x22').props()).to.have.property('value', '');
 
         await new Promise(resolve => setTimeout(resolve, 5));
 
@@ -292,7 +292,7 @@ describe('Everything', () => {
     });
 
     it('works (NumField, decimal)', async () => {
-        expect(wrapper.find('#x22').props()).to.have.property('value', 0);
+        expect(wrapper.find('#x22').props()).to.have.property('value', '');
         expect(wrapper.find('#x22').simulate('change', {target: {value: 2}})).to.be.ok;
         expect(wrapper.find('#x22').props()).to.have.property('value', 2);
 
@@ -417,7 +417,7 @@ describe('Everything', () => {
     it('works (rest)', () => {
         wrapper.setProps({grid:  10});
         wrapper.setProps({error: {}});
-        wrapper.setProps({model: {x09: ['', '', '']}});
+        wrapper.setProps({model: {...wrapper.state('model'), x09: ['', '', '']}});
 
         schema.x = {__type__: () => {}};