+++ packages/uniforms-unstyled/src/components/fields/NumField.js
@@ -36,7 +36,7 @@ const Num = ({
             ref={inputRef}
             step={decimal ? 0.01 : 1}
             type="number"
-            value={value === undefined ? null : value}
+            value={value}
         />
     </section>
 ;